+++ apps/demo-react/components/index.ts
@@ -1,7 +1,7 @@
 export { default as ConnectDisconnect } from './ConnectDisconnect';
 export { default as Header } from './Header';
 export { default as MainContainer } from './MainContainer';
-export { default as ProviderWeb3WithProps } from './ProviderWeb3WithProps';
+export { default as ProviderWeb3WithProps } from './ConfigContextProviders';
 export { default as ThemeSelect } from './ThemeSelect';
 export { default as WalletsModal } from './WalletsModal';
 export { default as ContractTesting } from './contractTesting/index';