+++ apps/demo-react/next.config.js
@@ -18,4 +18,26 @@ export default withBundleAnalyzer({
   publicRuntimeConfig: {
     alchemyApiKey,
   },
+  async headers() {
+    return [
+      {
+        // required for Gnosis Safe apps
+        source: '/manifest.json',
+        headers: [
+          {
+            key: "Access-Control-Allow-Origin",
+            value: '*',
+          },
+          {
+            key: "Access-Control-Allow-Methods",
+            value: 'GET',
+          },
+          {
+            key: "Access-Control-Allow-Headers",
+            value: 'X-Requested-With, content-type, Authorization',
+          },
+        ],
+      },
+    ]
+  }
 })