+++ packages/core-react/src/context/reefKnot.tsx
@@ -4,12 +4,14 @@ import { Chain } from 'wagmi/chains';
 import { WCWarnBannerContextProvider } from '@reef-knot/ui-react';
 import { getWalletDataList } from '../walletData/index';
 import { AcceptTermsModalContextProvider } from './acceptTermsModal';
+import { AutoConnect } from '../components/AutoConnect';
 
 export interface ReefKnotContextProps {
   rpc: Record<number, string>;
   walletconnectProjectId?: string;
   chains: Chain[];
   defaultChain: Chain;
+  autoConnect?: boolean;
   children?: ReactNode;
 }
 
@@ -26,6 +28,7 @@ export const ReefKnot: FC<ReefKnotContextProps> = ({
   walletconnectProjectId,
   chains,
   defaultChain,
+  autoConnect = true,
   children,
 }) => {
   const walletDataList = getWalletDataList({
@@ -47,7 +50,15 @@ export const ReefKnot: FC<ReefKnotContextProps> = ({
   return (
     <ReefKnotContext.Provider value={contextValue}>
       <AcceptTermsModalContextProvider>
-        <WCWarnBannerContextProvider>{children}</WCWarnBannerContextProvider>
+        <WCWarnBannerContextProvider>
+          <AutoConnect
+            autoConnect={autoConnect}
+            walletDataList={walletDataList}
+            chains={chains}
+          >
+            {children}
+          </AutoConnect>
+        </WCWarnBannerContextProvider>
       </AcceptTermsModalContextProvider>
     </ReefKnotContext.Provider>
   );