+++ packages/web3-react/src/helpers/index.ts
@@ -1,6 +1,4 @@
 export * from './providerDetectors';
-export * from './ua';
-export * from './openWindow';
-export { default as isUrl } from './isUrl';
+export * from './userAgents';
 export * from './interceptLedgerError';
 export * from './unsupportedChainError';