+++ packages/reef-knot/CHANGELOG.md
@@ -1,5 +1,17 @@
 # reef-knot
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [1046886]
+  - @reef-knot/ledger-connector@3.0.0
+  - @reef-knot/core-react@2.0.0
+  - @reef-knot/web3-react@2.0.0
+  - @reef-knot/connect-wallet-modal@2.0.0
+  - @reef-knot/wallets-list@1.12.0
+  - @reef-knot/types@1.5.0
+
 ## 1.15.3
 
 ### Patch Changes